Split add-ec2-tabpane and missing props
